Migrate from create-react-app to nextjs (#1130)
