+++ monitoring/ingest_inspector/frontend/.gitignore
@@ -1,24 +1,34 @@
-# Node modules
-node_modules/
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+.yarn/install-state.gz
 
-# Logs
-logs
-*.log
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+
+# production
+/build
+
+# misc
+.DS_Store
+*.pem
+
+# debug
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-lerna-debug.log*
-.pnpm-debug.log*
 
-# Optional npm cache directory
-.npm
+# local env files
+.env*.local
 
-# Yarn Integrity file
-.yarn-integrity
-.yarn/cache
-.yarn/unplugged
-.yarn/build-state.yml
-.yarn/install-state.gz
+# vercel
+.vercel
 
-# Build folders
-build/
+# typescript
+*.tsbuildinfo
+next-env.d.ts
