--- monitoring/ingest_inspector/frontend/.prettierignore
@@ -1,5 +1,0 @@
-node_modules
-
-# Ignore artifacts:
-build
-coverage
