--- monitoring/ingest_inspector/frontend/eslint.config.mjs
@@ -1,18 +1,0 @@
-import globals from "globals";
-import pluginJs from "@eslint/js";
-import tseslint from "typescript-eslint";
-import pluginReactConfig from "eslint-plugin-react/configs/recommended.js";
-
-export default [
-    {files: ["**/*.{js,mjs,cjs,ts,jsx,tsx}"]},
-    {languageOptions: {parserOptions: {ecmaFeatures: {jsx: true}}}},
-    {languageOptions: {globals: globals.browser}},
-    pluginJs.configs.recommended,
-    ...tseslint.configs.recommended,
-    pluginReactConfig,
-    {
-        rules: {
-            "react/react-in-jsx-scope": "off"
-        }
-    }
-];
