+++ monitoring/ingest_inspector/frontend/package.json
@@ -1,46 +1,29 @@
 {
-  "name": "react-cors-spa",
-  "version": "1.0.0",
-  "private": false,
-  "devDependencies": {
-    "@eslint/js": "^9.5.0",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.34.3",
-    "globals": "^15.6.0",
-    "prettier": "3.3.2",
-    "react-scripts": "^5.0.1",
-    "typescript-eslint": "^7.13.1"
+  "name": "wellcomecollection-ingest-inspector",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.14.6",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
     "classnames": "^2.5.1",
+    "next": "14.2.4",
     "nprogress": "^0.2.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.23.1",
-    "react-scripts": "5.0.1",
-    "swr": "^2.2.5",
-    "typescript": "^4.9.5"
+    "react": "^18",
+    "react-dom": "^18",
+    "swr": "^2.2.5"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.4",
+    "postcss": "^8",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5"
   }
 }
