--- monitoring/ingest_inspector/frontend/prettierrc.json
@@ -1,6 +1,0 @@
-{
-  "trailingComma": "es5",
-  "tabWidth": 4,
-  "semi": false,
-  "singleQuote": true
-}
