+++ monitoring/ingest_inspector/frontend/src/app/globals.css
@@ -1,3 +1,7 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 div.content {
   max-width: 950px;
   margin-left: auto;
@@ -371,4 +375,31 @@ strong {
   .status-failed .callback--processing.value {
     margin-bottom: calc(1.2em - 2px);
   }
+}
+
+#nprogress {
+  pointer-events: none;
+}
+
+#nprogress .bar {
+  background: #ffce3c;
+  position: fixed;
+  z-index: 1031;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 4px;
+}
+
+#nprogress .peg {
+  display: block;
+  position: absolute;
+  right: 0;
+  width: 100px;
+  height: 100%;
+  box-shadow:
+    0 0 10px #ffce3c,
+    0 0 5px #ffce3c;
+  opacity: 1;
+  transform: rotate(3deg) translate(0, -4px);
 }
