+++ monitoring/ingest_inspector/frontend/src/app/types.tsx
