+++ monitoring/ingest_inspector/frontend/src/app/utils.tsx
